+++ apps/app/src/test/scala/com/daml/network/integration/tests/DecentralizedSynchronizerSoftDomainMigrationIntegrationTest.scala
@@ -38,6 +38,7 @@ import com.daml.network.util.{
   ConfigScheduleUtil,
   Contract,
   ContractWithState,
+  UpdateHistoryComparator,
   WalletTestUtil,
 }
 import com.daml.network.validator.config.AppManagerConfig
@@ -56,7 +57,8 @@ import scala.jdk.OptionConverters.*
 class DecentralizedSynchronizerSoftDomainMigrationIntegrationTest
     extends SvIntegrationTestBase
     with ConfigScheduleUtil
-    with WalletTestUtil {
+    with WalletTestUtil
+    with UpdateHistoryComparator {
 
   // Fails with unexpected CreatedEvent roots
   override protected def runUpdateHistorySanityCheck: Boolean = false
@@ -113,6 +115,7 @@ class DecentralizedSynchronizerSoftDomainMigrationIntegrationTest
 
   "scheduled global domain upgrade happens" in { implicit env =>
     initDsoWithSv1Only() withClue "spin up Dso"
+    val ledgerBeginSv1 = sv1Backend.participantClient.ledger_api.state.end()
     val sv1WalletUser = onboardWalletUser(sv1WalletClient, sv1ValidatorBackend)
 
     val timeUntilNewRule = 5.seconds
@@ -701,6 +704,14 @@ class DecentralizedSynchronizerSoftDomainMigrationIntegrationTest
             spw.subscriptions.SubscriptionPayment.COMPANION,
           )
           map (c(_, sv1WalletUser)): _*
+      )
+    }
+
+    eventually() {
+      compareHistoryViaScanApi(
+        ledgerBeginSv1,
+        sv1Backend,
+        scancl("sv1ScanClient"),
       )
     }