+++ apps/common/src/test/scala/com/daml/network/store/StoreTest.scala
@@ -871,12 +871,14 @@ abstract class StoreTest extends AsyncWordSpec with BaseTest {
         exerciseResult: damlValue,
         offset: String = nextOffset(),
         txEffectiveAt: Instant = defaultEffectiveAt,
+        recordTime: Instant = defaultEffectiveAt,
     )(implicit store: HasIngestionSink): Future[TransactionTree] = {
       val tx = mkTx(
         offset,
         Seq(mkExercise(contract, interfaceId, choiceName, choiceArgument, exerciseResult)),
         domain,
         txEffectiveAt,
+        recordTime = recordTime,
       )
       store.testIngestionSink
         .ingestUpdate(