+++ apps/scan/src/main/scala/com/daml/network/scan/admin/http/ScanHttpEncodings.scala
@@ -4,49 +4,81 @@
 package com.daml.network.scan.admin.http
 
 import com.daml.ledger.javaapi.data.{CreatedEvent, ExercisedEvent, Identifier, TreeEvent}
-import com.daml.network.environment.ledger.api.{
-  LedgerClient,
-  ReassignmentUpdate,
-  TransactionTreeUpdate,
-}
+import com.daml.network.environment.ledger.api.ReassignmentEvent.{Assign, Unassign}
+import com.daml.network.environment.ledger.api.{ReassignmentUpdate, TransactionTreeUpdate}
 import com.daml.network.http.v0.definitions
 import com.daml.network.http.v0.definitions.UpdateHistoryItem
+import com.daml.network.store.TreeUpdateWithMigrationId
 import com.daml.network.util.{Contract, ValueJsonCodecCodegen}
 import com.digitalasset.canton.daml.lf.value.json.ApiCodecCompressed
 import com.digitalasset.canton.logging.ErrorLoggingContext
 
 import java.time.ZoneOffset
 import scala.jdk.CollectionConverters.*
+import scala.jdk.OptionConverters.*
 
 object ScanHttpEncodings {
 
   def ledgerTreeUpdateToHttp(
-      tx: LedgerClient.GetTreeUpdatesResponse,
-      migrationId: Long,
+      updateWithMigrationId: TreeUpdateWithMigrationId
   )(implicit elc: ErrorLoggingContext): UpdateHistoryItem = {
-    val tree = tx.update match {
-      case TransactionTreeUpdate(tree) => tree
-      case ReassignmentUpdate(_) =>
-        // TODO (#12552): deal with this case
-        throw new IllegalStateException("Got an unexpected transfer.")
-    }
 
-    definitions.UpdateHistoryItem.fromUpdateHistoryTransaction(
-      definitions
-        .UpdateHistoryTransaction(
-          tree.getUpdateId,
-          migrationId,
-          tree.getWorkflowId,
-          tree.getRecordTime.toString,
-          tx.domainId.toProtoPrimitive,
-          tree.getEffectiveAt.toString,
-          tree.getOffset,
-          tree.getRootEventIds.asScala.toVector,
-          tree.getEventsById.asScala.map { case (eventId, treeEvent) =>
-            eventId -> treeEventToHttp(treeEvent)
-          }.toMap,
+    updateWithMigrationId.update.update match {
+      case TransactionTreeUpdate(tree) =>
+        definitions.UpdateHistoryItem.fromUpdateHistoryTransaction(
+          definitions
+            .UpdateHistoryTransaction(
+              tree.getUpdateId,
+              updateWithMigrationId.migrationId,
+              tree.getWorkflowId,
+              tree.getRecordTime.toString,
+              updateWithMigrationId.update.domainId.toProtoPrimitive,
+              tree.getEffectiveAt.toString,
+              tree.getOffset,
+              tree.getRootEventIds.asScala.toVector,
+              tree.getEventsById.asScala.map { case (eventId, treeEvent) =>
+                eventId -> treeEventToHttp(treeEvent)
+              }.toMap,
+            )
         )
-    )
+      case ReassignmentUpdate(update) =>
+        update.event match {
+          case Assign(submitter, source, target, unassignId, createdEvent, counter) =>
+            definitions.UpdateHistoryItem.fromUpdateHistoryReassignment(
+              definitions.UpdateHistoryReassignment(
+                update.updateId,
+                update.offset.getOffset,
+                update.recordTime.toString,
+                definitions.UpdateHistoryAssignment(
+                  submitter.toProtoPrimitive,
+                  source.toProtoPrimitive,
+                  target.toProtoPrimitive,
+                  updateWithMigrationId.migrationId,
+                  unassignId,
+                  createdEventToHttp(createdEvent),
+                  counter,
+                ),
+              )
+            )
+          case Unassign(submitter, source, target, unassignId, contractId, counter) =>
+            definitions.UpdateHistoryItem.fromUpdateHistoryReassignment(
+              definitions.UpdateHistoryReassignment(
+                update.updateId,
+                update.offset.getOffset,
+                update.recordTime.toString,
+                definitions.UpdateHistoryUnassignment(
+                  submitter.toProtoPrimitive,
+                  source.toProtoPrimitive,
+                  updateWithMigrationId.migrationId,
+                  target.toProtoPrimitive,
+                  unassignId,
+                  counter,
+                  contractId.contractId,
+                ),
+              )
+            )
+        }
+    }
   }
 
   private def treeEventToHttp(treeEvent: TreeEvent)(implicit elc: ErrorLoggingContext) = {
@@ -69,6 +101,8 @@ object ScanHttpEncodings {
               event.getChildEventIds.asScala.toVector,
               encodeExerciseResult(event),
               event.isConsuming,
+              event.getActingParties.asScala.toVector,
+              event.getInterfaceId.map(templateIdString(_)).toScala,
             )
         )
       case _ =>
@@ -77,6 +111,11 @@ object ScanHttpEncodings {
   }
 
   def createdEventToHttp(event: CreatedEvent)(implicit elc: ErrorLoggingContext) = {
+    event.getContractKey.toScala.foreach { _ =>
+      throw new IllegalStateException(
+        "Contract keys are unexpected in UpdateHistory http encoded events"
+      )
+    }
     definitions
       .CreatedEvent(
         "created_event",
@@ -86,6 +125,8 @@ object ScanHttpEncodings {
         event.getPackageName,
         encodeContractPayload(event),
         event.getCreatedAt.atOffset(ZoneOffset.UTC),
+        event.getSignatories.asScala.toVector,
+        event.getObservers.asScala.toVector,
       )
   }