+++ docs/src/release_notes.rst
@@ -16,6 +16,11 @@ Upcoming
   * Added a new CometBFT Network Status dashboard that displays how much data is being exchanged with each peer on the CometBFT P2P network.
     This should should make it easier to diagnose connectivity problems between network peers.
 
+* Scan API
+
+  * Added the ``getUpdateById`` API in ``scan-internal.yaml``.
+    The ``getUpdateById`` API can be used to retrieve an update by its update ID.
+
 * UI
 
   * Gzip compression has been enabled for the Scan, Wallet, SV and CNS UIs.